==> notebooks/06_visualization.ipynb <==
import sys
from pathlib import Path

# Add project root to sys.path so `src/` can be imported
sys.path.append(str(Path("..").resolve()))


import pandas as pd
from src.plot_utils import plot_aggregated_time_series

# Load the tabular data
tabular_data = pd.read_parquet("../data/transformed/tabular_data.parquet")

# Separate features and target
features = tabular_data.drop(columns=["target"])
targets = tabular_data["target"]

# Plot any row, e.g. 1234
plot_aggregated_time_series(features, targets, row_id=50)
