==> notebooks/13_inference_pipeline.ipynb <==
import sys
from pathlib import Path

# Add project root to sys.path so `src/` can be imported
sys.path.append(str(Path("..").resolve()))


import pandas as pd
import joblib
import hopsworks
from datetime import datetime, timedelta
from pathlib import Path

from src.config import (
    HOPSWORKS_API_KEY,
    HOPSWORKS_PROJECT_NAME,
    FEATURE_VIEW_NAME,
    FEATURE_VIEW_VERSION,
    MODELS_DIR
)
from src.data_utils import transform_ts_data_info_features


from dotenv import load_dotenv

load_dotenv()

project = hopsworks.login(api_key_value=HOPSWORKS_API_KEY, project=HOPSWORKS_PROJECT_NAME)
fs = project.get_feature_store()
fv = fs.get_feature_view(name=FEATURE_VIEW_NAME, version=FEATURE_VIEW_VERSION)

predict_for_hour = pd.Timestamp.now(tz="UTC").ceil("h")
start_time = predict_for_hour - timedelta(days=28)
end_time = predict_for_hour - timedelta(hours=1)

ts_data = fv.get_batch_data(start_time=start_time, end_time=end_time)
ts_data = ts_data[ts_data["start_hour"].between(start_time, end_time)].sort_values(["start_station_id", "start_hour"]).reset_index(drop=True)
ts_data["start_hour"] = pd.to_datetime(ts_data["start_hour"])

ts_data.tail()


print(ts_data.shape)
